requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
frozenlist==1.7.0
idna==3.10
kombu==5.5.4
multidict==6.6.4
packaging==25.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
TgCrypto==1.2.5
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
yt-dlp==2025.8.22
